--- test/fixtures/unittest/fixture001/synctexjs/t.synctex
@@ -1,44 +1,0 @@
-SyncTeX Version:1
-Input:1:/Users/username/t.tex
-Input:2:/usr/local/texlive/2021/texmf-dist/tex/latex/base/article.cls
-Input:3:/usr/local/texlive/2021/texmf-dist/tex/latex/base/size12.clo
-Input:4:/usr/local/texlive/2021/texmf-dist/tex/latex/l3backend/l3backend-dvips.def
-Output:dvi
-Magnification:1000
-Unit:1
-X Offset:4736287
-Y Offset:4736287
-Content:
-!389
-{1
-[1,4:0,41451520:28114944,41451520,0
-(1,4:0,0:0,0,0
-)
-[1,4:2555904,41451520:25559040,40337408,0
-[1,4:2555904,1900544:25559040,786432,0
-h1,4:2555904,1900544:25559040,0,0
-]
-[1,4:2555904,39485440:25559040,35946496,0
-(1,4:2555904,4325376:25559040,546132,0
-h1,3:2555904,4325376:1155060,0,0
-x1,3:4523784,4325376
-x1,3:5315214,4325376
-k1,4:28114944,4325376:22799730
-g1,4:28114944,4325376
-)
-]
-(1,4:2555904,41451520:25559040,506811,0
-k1,4:15142914,41451520:12587010
-x1,4:15527934,41451520
-k1,4:28114944,41451520:12587010
-)
-]
-]
-!524
-}1
-Input:5:/Users/username/./t.aux
-!79
-Postamble:
-Count:25
-!24
-Post scriptum:
