--- test/fixtures/unittest/fixture001/textdocumentlike/abc.txt
@@ -1,7 +1,0 @@
-abc
-d
-
-efg
-
-hijk
-lmn
