--- test/fixtures/unittest/fixture020_structure/.vscode/settings.json
@@ -1,2 +1,0 @@
-{
-}
