--- test/fixtures/unittest/fixture020_structure/main.tex
@@ -1,49 +1,0 @@
-\documentclass{article}
-\usepackage{import}
-\usepackage{hyperref}
-
-\begin{document}
-
-\input{sub_1}
-
-\label{sec11}
-
-\subsubsection{1.1.1}
-
-\subsection{1.2}
-
-\import{subs/}{sub2.tex}
-
-\subsubsection{2.0.1}
-
-\subimport{subs/}{sub3.tex}
-
-\section{4 A long title split
-over two lines}
-
-\section*{No \textit{Number} Section}
-
-\section{Section \texorpdfstring{tex}{pdf} Caption}
-
-\begin{figure}
-\end{figure}
-
-\begin{figure}
-\caption{Figure Caption}
-\end{figure}
-
-\begin{table}
-\caption{Table Caption}
-\end{table}
-
-\begin{frame}
-\frametitle{Frame Title 1}
-\end{frame}
-
-\begin{frame} {Frame Title 2}
-\end{frame}
-
-\begin{frame}
-\end{frame}
-
-\end{document}
