--- test/fixtures/unittest/fixture020_structure/sub_1.tex
@@ -1,5 +1,0 @@
-\section{1}\label{sec1}
-
-\subsection{1.1}
-
-\altsection{1.1?}
