--- test/fixtures/unittest/fixture020_structure/subs/sub2.tex
@@ -1,3 +1,0 @@
-\subsubsection{1.2.1}
-
-\section{2}
