--- test/fixtures/unittest/fixture020_structure/subs/sub3.tex
@@ -1,1 +1,0 @@
-\section{3}
