--- test/fixtures/unittest/fixture030_linter/chktex.linterlog
@@ -1,2 +1,0 @@
-main.tex:5:18:1:Warning:24:Delete this space to maintain correct pagereferences.
-sub/sub.tex:1:26:1:Warning:24:Delete this space to maintain correct pagereferences.
