--- test/fixtures/unittest/fixture030_linter/lacheck.linterlog
@@ -1,3 +1,0 @@
-"main.tex", line 7: double space at "~~"
-** sub/sub:
-"sub/sub.tex", line 2: double space at "~~"
