--- test/fixtures/unittest/fixture030_linter/main.tex
@@ -1,10 +1,0 @@
-\documentclass{article}
-
-\begin{document}
-
-\section{Section} \label{section}
-
-LaCheck~~Test
-\input{sub/sub}
-
-\end{document}
