--- test/fixtures/unittest/fixture030_linter/sub/sub.tex
@@ -1,2 +1,0 @@
-\section{Another Section} \label{another}
-LaCheck~~Sub
