--- test/unittests/providers/completer/commandlib/commandfinder.test.ts
@@ -1,12 +1,0 @@
-import {runUnitTestWithFixture} from '../../../../utils/ciutils'
-import {isTriggerSuggestNeeded} from '../../../../../src/providers/completer/commandlib/commandfinder'
-import assert from 'assert'
-
-
-suite('unit test suite: commandlib/commandfinder', () => {
-
-    runUnitTestWithFixture('fixture001', 'test isTriggerSuggestNeeded', () => {
-        assert.ok(!isTriggerSuggestNeeded('frac'))
-    })
-
-})
